cli: reorganize tasks
