+++ cli/commands/contracts/ens.ts
@@ -7,7 +7,7 @@ import { logger } from '../../logging'
 
 export const registerTestName = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const name = cliArgs.name
-  const testRegistrar = cli.contracts.ITestRegistrar
+  const testRegistrar = (cli.contracts as any).ITestRegistrar
   const normalizedName = name.toLowerCase()
   const labelNameFull = `${normalizedName}.${'eth'}`
   const labelHashFull = utils.namehash(labelNameFull)
