+++ cli/commands/contracts/gsr-gdai.ts
@@ -7,7 +7,7 @@ import { loadEnv, CLIArgs, CLIEnvironment } from '../../env'
 
 export const setGSR = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const account = cliArgs.account
-  const gdai = cli.contracts.GDAI
+  const gdai = (cli.contracts as any).GDAI
 
   logger.log(`Setting GSR to ${account}...`)
   await sendTransaction(cli.wallet, gdai, 'setGSR', [account])
@@ -15,7 +15,7 @@ export const setGSR = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<voi
 
 export const setRate = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
-  const gsr = cli.contracts.GSRManager
+  const gsr = (cli.contracts as any).GSRManager
 
   logger.log(`Setting rate to ${amount}...`)
   await sendTransaction(cli.wallet, gsr, 'setRate', [amount])
@@ -23,7 +23,7 @@ export const setRate = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<vo
 
 export const join = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
-  const gsr = cli.contracts.GSRManager
+  const gsr = (cli.contracts as any).GSRManager
 
   logger.log(`Reminder - you must call approve on the GSR before`)
   logger.log(`Joining GSR with ${cliArgs.amount} tokens...`)
@@ -33,7 +33,7 @@ export const join = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void>
 export const mint = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
   const account = cliArgs.account
-  const gdai = cli.contracts.GDAI
+  const gdai = (cli.contracts as any).GDAI
 
   logger.log(`Minting ${cliArgs.amount} GDAI for user ${account}...`)
   await sendTransaction(cli.wallet, gdai, 'mint', [account, amount])
@@ -41,7 +41,7 @@ export const mint = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void>
 
 export const burn = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
-  const gdai = cli.contracts.GDAI
+  const gdai = (cli.contracts as any).GDAI
 
   logger.log(`Burning ${cliArgs.amount} GDAI...`)
   await sendTransaction(cli.wallet, gdai, 'burn', [amount])
@@ -50,7 +50,7 @@ export const burn = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void>
 export const transfer = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
   const account = cliArgs.account
-  const gdai = cli.contracts.GDAI
+  const gdai = (cli.contracts as any).GDAI
 
   logger.log(`Transferring ${cliArgs.amount} tokens to user ${account}...`)
   await sendTransaction(cli.wallet, gdai, 'transfer', [account, amount])
@@ -59,7 +59,7 @@ export const transfer = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<v
 export const approve = async (cli: CLIEnvironment, cliArgs: CLIArgs): Promise<void> => {
   const amount = parseGRT(cliArgs.amount)
   const account = cliArgs.account
-  const gdai = cli.contracts.GDAI
+  const gdai = (cli.contracts as any).GDAI
 
   logger.log(`Approving ${cliArgs.amount} GDAI for user ${account} to spend...`)
   await sendTransaction(cli.wallet, gdai, 'approve', [account, amount])
